20-joule-a2a-code-based-agent/deep_research_api/app/requirements.txt
# Web framework & server
fastapi>=0.115.0
uvicorn[standard]>=0.34.2

# Deep Agents SDK
deepagents>=0.2.6

# SAP Generative AI Hub (includes gen_ai_hub.proxy.langchain.init_llm)
sap-ai-sdk-gen[all]

# LangChain / LangGraph
langgraph>=0.3.18
langchain-core>=0.3.0

# Research tools
tavily-python>=0.5.0
markdownify>=1.2.0

# Utilities
python-dotenv>=1.1.0
pydantic>=2.10.6
